requirements.txt
jira==3.3.1
pydantic==1.9.1
python-dotenv==0.20.0
